Устранил баги с менеджментом проектов, с добавлением, удалением и подсчетом количества. Сделал первую версию подключения Yclients.

+++ src/main/java/com/kuzmichev/AdMetricsBot/model/Project.java
@@ -15,5 +15,5 @@ public class Project {
     private String projectId;
     private String projectName;
     private String chatId;
-    private int projectNumber;
+    private long projectNumber;
 }
